Fixed some issues. Actually don't know what they are...

+++ src/datatypes/utf8.c
@@ -70,7 +70,7 @@ bool utf8_eq(Utf8 s1, Utf8 s2) {
 }
 
 bool utf8_try_next(Utf8 *str, Utf8 *out) {
-  usize bit_count = u8_leading_ones(str->data[0]);
+  usize bit_count = u8_leading_ones((u8)str->data[0]);
   clib_assert_return(bit_count <= 4, false);
   clib_assert_return(bit_count != 1, false);
   usize bytes = usize_max(1, bit_count);
